refactoring : reconstruct folder structure
