+++ app/src/components/Home/BarChart.jsx
