+++ app/src/components/Home/ChartContainer.jsx
@@ -1,7 +1,7 @@
-import BarChart from "../components/BarChart";
-import Loading from "../components/Loading";
+import BarChart from "@components/Home/BarChart";
+import Loading from "@components/Loading";
 import React from "react";
-import axisName from "../utils/axisName";
+import axisName from "@utils/axisName";
 import { useSelector } from "react-redux";
 
 /**
