+++ app/src/components/Modal/AlertModal.jsx
@@ -1,5 +1,5 @@
 import { Button } from "@material-ui/core";
-import Modal from "./Modal";
+import Modal from "@components/Modal";
 import React from "react";
 
 function AlertModal({ closeModal, children }) {
