+++ app/webpack.config.js
@@ -10,17 +10,30 @@ dotenv.config();
 module.exports = {
   mode: "development",
   entry: {
-    app: "./src/index.js",
+    app: "./src/index.jsx",
   },
   output: {
     path: path.resolve(__dirname, "dist"),
     filename: "index_bundler.js",
   },
   node: { fs: "empty" },
+  resolve: {
+    extensions: [".js", ".jsx"],
+    alias: {
+      "@api": path.resolve(__dirname, "./src/api"),
+      "@assets": path.resolve(__dirname, "./src/assets"),
+      "@components": path.resolve(__dirname, "./src/components"),
+      "@hooks": path.resolve(__dirname, "./src/hooks"),
+      "@layouts": path.resolve(__dirname, "./src/layouts"),
+      "@modules": path.resolve(__dirname, "./src/modules"),
+      "@pages": path.resolve(__dirname, "./src/pages"),
+      "@utils": path.resolve(__dirname, "./src/utils"),
+    },
+  },
   module: {
     rules: [
       {
-        test: /\.js$/,
+        test: /\.jsx$/,
         exclude: [/node_modules/],
         use: [
           {
